requirements.txt
aiohttp==3.9.1
asyncio==3.4.3
ccxt==4.2.25
openai==1.6.1
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0
azure-storage-blob==12.19.0
azure-data-tables==12.4.4
azure-storage-file-datalake==12.14.0
azure-monitor-opentelemetry==1.2.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-aiohttp-client==0.43b0
opentelemetry-exporter-otlp==1.22.0
opencensus-ext-azure==1.1.13
pyodbc==5.0.1
pandas==2.1.4
numpy==1.26.2
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
httpx==0.26.0
tenacity==8.2.3
black==23.12.1
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

